% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_graphab_object.R
\name{check_graphab_object}
\alias{check_graphab_object}
\title{Check whether a Graphab object exists in the Graphab project.}
\usage{
check_graphab_object(proj_path, object_type, name)
}
\arguments{
\item{proj_path}{The path to the project .xml file.}

\item{object_type}{A character string with the name of the object type whose
list will be checked. Can be either "habitat", "linkset", or "graph".}

\item{name}{A character string (or vector) with the name(s) of the
object(s) whose existence is checked.}
}
\value{
A logical specifying whether the object exists.
}
\description{
The function checks whether a given object exists in the
Graphab project. This can be a habitat, a linkset or a graph.
}
\examples{
\dontrun{
proj_end_path <- "./graphab_example/graphab_example.xml"
check_graphab_object(proj_path = proj_end_path,
                     object_type = "habitat",
                     name = "forest")
}
}
\author{
P. Savary
}
\keyword{internal}
