% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_graphab_version.R
\name{check_graphab_version}
\alias{check_graphab_version}
\title{Check whether the Graphab project has been created with graphab.jar >= 3.0}
\usage{
check_graphab_version(proj_path)
}
\arguments{
\item{proj_path}{The path to the project .xml file.}
}
\value{
If the version is < graphab-3.0.jar, the function returns a
message specifying that the project has been created with an earlier
version and that this version of graph4lg should not be used, or that
another project should be created.
}
\description{
The function checks whether the Graphab project has been
created with graphab.jar >= 3.0
}
\examples{
\dontrun{
proj_end_path <- "./graphab_example/graphab_example.xml"
check_graphab_version(proj_path = proj_end_path)
}
}
\author{
P. Savary
}
\keyword{internal}
