% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ex_gstud}
\alias{data_ex_gstud}
\title{data_ex_gstud genetic dataset}
\format{
A 'data.frame' with columns:
\describe{
\item{ID}{Individual ID}
\item{POP}{Population name}
\item{LOCI-1 to LOCI-20}{20 loci columns with microsatellite data with
3 digits coding, alleles separated by ":", and blank missing data
(class 'locus' from \pkg{gstudio})}
}
}
\usage{
data_ex_gstud
}
\description{
Genetic dataset from genetic simulation on CDPOP
200 individuals, 10 populations
20 microsatellite loci (3 digits coding)
100 generations simulated
}
\examples{
data("data_ex_gstud")
str(data_ex_gstud)
length(unique(data_ex_gstud$POP))
}
\keyword{datasets}
