% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_ex_loci}
\alias{data_ex_loci}
\title{data_ex_loci genetic dataset}
\format{
An object of class 'loci' and 'data.frame' with the columns :
\describe{
\item{population}{Population name}
\item{Other columns}{20 loci columns with microsatellite data with
3 digits coding, alleles separated by "/", and missing data noted "NA/NA"}
}
Row names correspond to individuals' ID
}
\usage{
data_ex_loci
}
\description{
Genetic dataset from genetic simulation on CDPOP
200 individuals, 10 populations
20 microsatellite loci (3 digits coding)
100 generations simulated
}
\examples{
data("data_ex_loci")
length(unique(data_ex_loci$population))
}
\keyword{datasets}
