% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pc_genind}
\alias{data_pc_genind}
\title{data_pc_genind genetic dataset}
\format{An object of type 'genind'}
\source{
\url{https://datadryad.org/resource/doi:10.5061/dryad.v8324}
}
\usage{
data_pc_genind
}
\description{
Setophaga plumbea genetic dataset
Data collected in French Guadeloupe by A. Khimoun and S. Garnier
432 individuals, 20 populations
12 microsatellite loci (3 digits coding)
169 different alleles
}
\examples{
data("data_pc_genind")
length(unique(data_pc_genind@pop))
}
\references{
\insertRef{khimoun2017landscape}{graph4lg}
}
\keyword{datasets}
