% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pc_gpop}
\alias{data_pc_gpop}
\title{data_pc_gpop genetic dataset}
\format{A 'data.frame' with the format needed to save as a GENEPOP software
file :
\describe{
\item{1st column}{Individual identifier beginning with population name and
ending with a comma}
\item{Other columns}{12 loci's columns with microsatellites' data with
3 digits coding, alleles not separated, and missing data noted "000000"}
}
Each group of individuals from the same population are separated by a line
with only the character value "Pop" in firts column.
First lines include loci names and file name}
\source{
\url{https://datadryad.org/resource/doi:10.5061/dryad.v8324}
}
\usage{
data_pc_gpop
}
\description{
Setophaga plumbea genetic dataset
Data collected in French Guadeloupe by A. Khimoun and S. Garnier
432 individuals, 20 populations
12 microsatellite loci (3 digits coding)
169 different alleles
}
\examples{
data("data_pc_gpop")
str(data_pc_gpop)
}
\references{
\insertRef{khimoun2017landscape}{graph4lg}
}
\keyword{datasets}
