% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pc_gstud}
\alias{data_pc_gstud}
\title{data_pc_gstud genetic dataset}
\format{A 'data.frame' with columns:
\describe{
\item{Species}{Species identifier: in this case 'PC'}
\item{Cluster}{Population name}
\item{ID}{Individual ID}
\item{id}{Population ID used during the sampling stage}
\item{Latitude}{Sampling site (population) latitude (WGS 84 - UTM 20 N)}
\item{Longitude}{Sampling site (population) longitude (WGS 84 - UTM 20 N)}
\item{DkiD102 to DkiD12}{12 loci columns with microsatellite data with
3 digits coding, alleles separated by ":", and blank missing data
(class 'locus' from \pkg{gstudio})}
}}
\source{
\url{https://datadryad.org/resource/doi:10.5061/dryad.v8324}
}
\usage{
data_pc_gstud
}
\description{
Setophaga plumbea genetic dataset
Data collected in French Guadeloupe by A. Khimoun and S. Garnier
432 individuals, 20 populations
12 microsatellite loci (3 digits coding)
169 different alleles
}
\examples{
data("data_pc_gstud")
str(data_pc_gstud)
length(unique(data_pc_gstud$Cluster))
}
\references{
\insertRef{khimoun2017landscape}{graph4lg}
}
\keyword{datasets}
