% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pc_loci}
\alias{data_pc_loci}
\title{data_pc_loci genetic dataset}
\format{An object of class 'loci' and 'data.frame' with the columns :
\describe{
\item{population}{Population name}
\item{Other columns}{12 loci columns with microsatellite data with
3 digits coding, alleles separated by "/", and missing data noted "NA/NA"}
}
Row names correspond to individuals' ID}
\source{
\url{https://datadryad.org/resource/doi:10.5061/dryad.v8324}
}
\usage{
data_pc_loci
}
\description{
Setophaga plumbea genetic dataset
Data collected in French Guadeloupe by A. Khimoun and S. Garnier
432 individuals, 20 populations
12 microsatellite loci (3 digits coding)
169 different alleles
}
\examples{
data("data_pc_loci")
length(unique(data_pc_loci$population))
}
\references{
\insertRef{khimoun2017landscape}{graph4lg}
}
\keyword{datasets}
