% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pc_str}
\alias{data_pc_str}
\title{data_pc_str genetic dataset}
\format{An object of class 'data.frame' ready to be saved as a STRUCTRE
software file. The resulting text file would include the following columns :
\describe{
\item{1st column}{Individual names}
\item{2nd column}{Population names}
\item{Other columns}{12 loci's columns with microsatellites' data with
3 digits coding, and missing data noted "-9"}
}
Each individual is displayed on two rows, with one allele on each row.
There are as many rows as twice the number of individuals.}
\source{
\url{https://datadryad.org/resource/doi:10.5061/dryad.v8324}
}
\usage{
data_pc_str
}
\description{
Setophaga plumbea genetic dataset
Data collected in French Guadeloupe by A. Khimoun and S. Garnier
432 individuals, 20 populations
12 microsatellite loci (3 digits coding)
169 different alleles
}
\examples{
data("data_pc_str")
}
\references{
\insertRef{khimoun2017landscape}{graph4lg}
}
\keyword{datasets}
