% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_simul_genind}
\alias{data_simul_genind}
\title{data_simul_genind genetic dataset}
\format{
An object of type 'genind'
}
\usage{
data_simul_genind
}
\description{
Genetic dataset from genetic simulation on CDPOP
1500 individuals, 50 populations
20 microsatellite loci (3 digits coding)
50 generations simulated
}
\details{
The simulation was made with CDPOP during 50 generations.
Dispersal was possible between the 50 populations. Its probability depended
on the cost distance between populations, calculated on a simulated
resistance surface (raster). Mutations were not possible. There
were initially 600 alleles in total (many disappeared because of drift).
Population stayed constant
with a sex-ratio of 1. Generations did not overlap.
This simulation includes a part of stochasticity and these data result
from only 1 simulation run.
}
\examples{
data("data_simul_genind")
length(unique(data_simul_genind@pop))
}
\references{
\insertRef{landguth2010cdpop}{graph4lg}
}
\keyword{datasets}
