% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_tuto}
\alias{data_tuto}
\title{data_tuto : data used to generate the vignette}
\format{
Several outputs or inputs to show how the package works in a list
\describe{
\item{mat_dps}{Genetic distance matrix example}
\item{mat_pg}{Second genetic distance matrix example}
\item{graph_ci}{Genetic independence graph example}
\item{dmc}{Output of the function 'dist_max_corr'}
\item{land_graph}{Landscape graph example}
\item{mat_ld}{Landscape distance matrix example}
}

Several outputs or inputs to show how the package works in a list
\describe{
\item{dmc}{Output of the function 'dist_max_corr'}
\item{graph_ci}{Genetic independence graph example}
\item{mat_dps}{Genetic distance matrix example}
\item{mat_pg}{Second genetic distance matrix example}
}
}
\usage{
data_tuto

data_tuto
}
\description{
Data used to generate the vignette

Data used to generate the vignette
}
\examples{
data("data_tuto")
mat_dps <- data_tuto[[1]]
str(mat_dps)
data("data_tuto")
mat_dps <- data_tuto[[1]]
str(mat_dps)
}
\keyword{datasets}
