% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_great_circle.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{A coordinate in degrees}
}
\value{
The coordinate in radians
}
\description{
The function converts degree to radians
}
\examples{
deg2rad(40.75170)
}
\author{
P. Savary
}
\keyword{internal}
