% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_great_circle.R
\name{dist_great_circle}
\alias{dist_great_circle}
\title{Compute the Great Circle distance between two points}
\usage{
dist_great_circle(long1, long2, lat1, lat2, method = "vicenty")
}
\arguments{
\item{long1}{project name, project dir in which proj_name.xml is found}

\item{long2}{raster.tif INT2S path or present in wd,}

\item{lat1}{habitat code in the raster file}

\item{lat2}{default 0, minimum habitat size in ha}

\item{method}{default NULL nodata code in the raster file}
}
\description{
The function computes the Great Circle distance between two
two points defined by their longitudes and latitudes.
}
\examples{
dist_great_circle(long1 = -73.99420,
                  lat1 = 40.75170,
                  long2 = -87.63940,
                  lat2 = 41.87440,
                  method = "vicenty")
}
\author{
P. Savary
}
\keyword{internal}
