% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphab_merge_graph.R
\name{graphab_merge_graph}
\alias{graphab_merge_graph}
\title{Merge graphs in a multiple-graph Graphab project}
\usage{
graphab_merge_graph(
  proj_name,
  graphs,
  name = NULL,
  proj_path = NULL,
  parallel.java = NULL,
  alloc_ram = NULL
)
}
\arguments{
\item{proj_name}{A character string indicating the Graphab project name.
The project name is also the name of the project directory in which the
file proj_name.xml is. It can be created with \code{\link{graphab_project}}}

\item{graphs}{A character vector indicating the names of the graphs to be
merged. These graphs must already exist in the project. Graphs can be
created with \code{\link{graphab_graph}}.}

\item{name}{A character string indicating the name of the graph created
by merging the other graphs.}

\item{proj_path}{(optional) A character string indicating the path to the
directory that contains the project directory. It should be used when the
project directory is not in the current working directory. Default is NULL.
When 'proj_path = NULL', the project directory is equal to \code{getwd()}.}

\item{parallel.java}{An integer indicating how many computer cores are used
to run the .jar file. By default, \code{parallel.java = NULL}, and java sets
it according to local settings.}

\item{alloc_ram}{(optional, default = NULL) Integer or numeric value
indicating RAM gigabytes allocated to the java process. Increasing this
value can speed up the computations. Too large values may not be compatible
with your machine settings.}
}
\description{
The function creates one graph from several other graphs
}
\details{
By default, intra-patch distances are considered for metric
calculation. See more information in Graphab 3.0 manual:
\url{https://thema.umlp.fr/productions/software/graphab/download/manual-3.0-en.pdf}
}
\examples{
\dontrun{
graphab_merge_graph(proj_name = "graphab_example",
              graphs = c("graph_forest_1", "graph_shrub_2"),
              name = "g_all")
}
}
\author{
P. Savary
}
