% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphab_metapatch.R
\name{graphab_metapatch}
\alias{graphab_metapatch}
\title{Create a new habitat type made of meta-patches based on inter-patch
distances in an existing graph}
\usage{
graphab_metapatch(
  proj_name,
  graph,
  mincapa = NULL,
  proj_path = NULL,
  parallel.java = NULL,
  alloc_ram = NULL
)
}
\arguments{
\item{proj_name}{A character string indicating the Graphab project name.
The project name is also the name of the project directory in which the
file proj_name.xml is. It can be created with \code{\link{graphab_project}}}

\item{graph}{A character string indicating the name of the graph from which
the meta-patches are created. This graph should have several components. Each
of them includes one or several habitat patches and they will each correspond
to a single habitat (meta-)patch in the new habitat type created by this
function. The graph should have been created using the
\code{\link{graphab_graph}} function and is associated with a link set. The
latter should not be complete.}

\item{mincapa}{An integer or numeric value indicating the minimum
capacity of the meta-patch to conserve in the new habitat type. This capacity
is computed as the sum of the capacities of all the patches included in a
meta-patch.}

\item{proj_path}{(optional) A character string indicating the path to the
directory that contains the project directory. It should be used when the
project directory is not in the current working directory. Default is NULL.
When 'proj_path = NULL', the project directory is equal to \code{getwd()}.}

\item{parallel.java}{An integer indicating how many computer cores are used
to run the .jar file. By default, \code{parallel.java = NULL}, and java sets
it according to local settings.}

\item{alloc_ram}{(optional, default = NULL) Integer or numeric value
indicating RAM gigabytes allocated to the java process. Increasing this
value can speed up the computations. Too large values may not be compatible
with your machine settings.}
}
\description{
The function creates a new habitat type made of meta-patches
based on inter-patch distances in an existing pruned graph
}
\details{
See more information in Graphab 3.0 manual:
\url{https://thema.umlp.fr/productions/software/graphab/download/manual-3.0-en.pdf}
Note that a name is automatically given to this new habitat type.
}
\examples{
\dontrun{
graphab_metapatch(proj_name = "graphab_example",
                 graph = "graph_forest_thr200")
}
}
\references{
\insertRef{foltete2012software}{graph4lg}
\insertRef{foltete2021graphab}{graph4lg}
\insertRef{savary2024multiple}{graph4lg}
}
\author{
P. Savary
}
