% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphab_pointset.R
\name{graphab_pointset}
\alias{graphab_pointset}
\title{Add a point set to the Graphab project}
\usage{
graphab_pointset(
  proj_name,
  linkset = NULL,
  name,
  pointset,
  id = NULL,
  return_val = NULL,
  proj_path = NULL,
  parallel.java = NULL,
  alloc_ram = NULL
)
}
\arguments{
\item{proj_name}{A character string indicating the Graphab project name.
The project name is also the name of the project directory in which the
file proj_name.xml is.}

\item{linkset}{Deprecated parameter from graph4lg < 2.0}

\item{name}{A character string indicating the name of the new habitat type
created by adding this point set to the project.}

\item{pointset}{Can be either:\itemize{
\item{A character string indicating the path (absolute or relative) to a
geopackage point layer with its extension (.gpkg).}
\item{A character string indicating the path to a .csv file with three
columns: ID, x and y, respectively indicating the point ID, longitude
and latitude.}
\item{A data.frame with three columns:
ID, x and y, respectively indicating the point ID, longitude and latitude.}
}}

\item{id}{Deprecated parameter from graph4lg < 2.0}

\item{return_val}{Deprecated parameter from graph4lg < 2.0}

\item{proj_path}{(optional) A character string indicating the path to the
directory that contains the project directory. It should be used when the
project directory is not in the current working directory. Default is NULL.
When 'proj_path = NULL', the project directory is equal to \code{getwd()}.}

\item{parallel.java}{An integer indicating how many computer cores are used
to run the .jar file. By default, \code{parallel.java = NULL}, and java sets
it according to local settings.}

\item{alloc_ram}{(optional, default = NULL) Integer or numeric value
indicating RAM gigabytes allocated to the java process. Increasing this
value can speed up the computations. Too large values may not be compatible
with your machine settings.}
}
\description{
The function adds a spatial point set to the Graphab project.
These points will define a new habitat type with 0-capacity at first.
}
\details{
Point coordinates must be in the same coordinate reference system
as the habitat patches (and initial raster layer). See more information in
Graphab 3.0 manual:
\url{https://thema.umlp.fr/productions/software/graphab/download/manual-3.0-en.pdf}
}
\examples{
\dontrun{
graphab_pointset(proj_name = "graphab_example",
                 name = "pts_pop",
                 pointset = "pts_pop.gpkg")
}
}
\references{
\insertRef{foltete2012software}{graph4lg}
\insertRef{foltete2021graphab}{graph4lg}
\insertRef{savary2024multiple}{graph4lg}
}
\author{
P. Savary
}
