% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphab_project.R
\name{graphab_project}
\alias{graphab_project}
\title{Create a Graphab project}
\usage{
graphab_project(
  proj_name,
  raster,
  nodata = NULL,
  parallel.java = NULL,
  alloc_ram = NULL,
  proj_path = NULL
)
}
\arguments{
\item{proj_name}{A character string indicating the Graphab project name.
The project name is also the name of the project directory in which the
file proj_name.xml will be created.}

\item{raster}{A character string indicating the name of the .tif raster file
or of its path. If the path is not specified, the raster must be present in
the current working directory. Raster cell values must be in INT2S encoding.}

\item{nodata}{(optional, default=NULL) An integer or numeric value
specifying the code in the raster file associated with nodata value
(often corresponding to peripheric cells)}

\item{parallel.java}{An integer indicating how many computer cores are used
to run the .jar file. By default, \code{parallel.java = NULL}, and java sets
it according to local settings.}

\item{alloc_ram}{(optional, default = NULL) Integer or numeric value
indicating RAM gigabytes allocated to the java process. Increasing this
value can speed up the computations. Too large values may not be compatible
with your machine settings.}

\item{proj_path}{(optional) A character string indicating the path to the
directory that contains the project directory. It should be used when the
project directory is not in the current working directory. Default is NULL.
When 'proj_path = NULL', the project directory is equal to \code{getwd()}.}
}
\description{
The function creates a Graphab project from a raster file on
which habitat patches can be delimited.
}
\details{
A habitat patch consists of the central pixel with its eight
neighbors if they are of the same value (8-connexity) and the path
geometry is not simplified. See more information in Graphab 3.0 manual:
\url{https://thema.umlp.fr/productions/software/graphab/download/manual-3.0-en.pdf}
}
\examples{
\dontrun{
proj_name <- "graphab_example"
raster <- "rast_simul50.tif"
graphab_project(proj_name = proj_name,
               raster = raster)
}
}
\references{
\insertRef{foltete2012software}{graph4lg}
\insertRef{foltete2021graphab}{graph4lg}
\insertRef{savary2024multiple}{graph4lg}
}
\author{
P. Savary, T. Rudolph
}
