% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphab_project_desc.R
\name{graphab_project_desc}
\alias{graphab_project_desc}
\title{Describe the objects of a Graphab project}
\usage{
graphab_project_desc(proj_name, proj_path = NULL)
}
\arguments{
\item{proj_name}{A character string indicating the Graphab project name.
The project name is also the name of the project directory in which the
file proj_name.xml has been created.}

\item{proj_path}{(optional) A character string indicating the path to the
directory that contains the project directory. It should be used when the
project directory is not in the current working directory. Default is NULL.
When 'proj_path = NULL', the project directory is equal to \code{getwd()}.}
}
\value{
It returns a large list describing the objects in several categories,
including:\itemize{
\item{Source raster: the values of the source raster file.}
\item{Habitats: the list of created habitats, their code, their name,
their type (raster, vector), and, if raster, their code in the raster.}
\item{Linksets: the list of created linksets and their type. If computed
in cost distance units, a table of costs or the name of the external cost
raster file is returned.}
\item{Graphs: the list of graphs that have been created.}
\item{Metrics: the list of computed metrics and their associated graphs.}
}
}
\description{
The function describes the objects of a Graphab project
}
\examples{
\dontrun{
graphab_project_desc(proj_name = "graphab_example")
}
}
\author{
P. Savary
}
