% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphab_show.R
\name{graphab_show}
\alias{graphab_show}
\title{Show the objects created in the Graphab project}
\usage{
graphab_show(proj_path)
}
\arguments{
\item{proj_path}{The path to the project .xml file.}
}
\value{
The list of objects

A list of the existing objects, with NULL values if none exists.
}
\description{
The function lists the objects created in the project.
}
\details{
Given that this function calls Java and is heavier than other
check-ups, functions reading the .xml file are usually preferred.
}
\examples{
\dontrun{
proj_end_path <- "./graphab_example/graphab_example.xml"
graphab_show(proj_path = proj_end_path)
}
}
\author{
P. Savary
}
\keyword{internal}
