% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{harm_mean}
\alias{harm_mean}
\title{Compute the harmonic mean of a numeric vector}
\usage{
harm_mean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric value corresponding to the harmonic mean of the vector
}
\description{
The function computes the harmonic mean of a numeric vector
}
\examples{
x <- c(10, 2, 5, 15)
hm <- harm_mean(x)
}
\author{
P. Savary
}
\keyword{internal}
