% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_objects.R
\docType{data}
\name{mypalette}
\alias{mypalette}
\title{Vector of custom colors}
\format{
An object of class \code{character} of length 27.
}
\usage{
mypalette
}
\description{
Vector of custom colors
}
\examples{
mypalette[1]
}
\keyword{internal}
