% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pts_pop_simul}
\alias{pts_pop_simul}
\title{pts_pop_simul : details on simulated populations}
\format{
An object of class 'data.frame' with the following columns :
\describe{
\item{ID}{Population ID of the 50 populations}
\item{x}{Site longitude (RGF93)}
\item{y}{Site latitude (RGF93)}
}
}
\usage{
pts_pop_simul
}
\description{
Simulation dataset
50 populations located on a simulated landscape
}
\examples{
data("pts_pop_simul")
str(pts_pop_simul)
}
\references{
\insertRef{landguth2010cdpop}{graph4lg}
There are as many rows as there are sampled populations.
}
\keyword{datasets}
