% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_objects.R
\name{sc01}
\alias{sc01}
\title{Scaling function}
\usage{
sc01(x)
}
\arguments{
\item{x}{Numeric or integer vector}
}
\description{
Scales values between 0 and 1
}
\examples{
x <- runif(min = 3, max = 15, n = 20)
x01 <- sc01(x)
}
\keyword{internal}
