% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_mat_to_df.R
\name{pw_mat_to_df}
\alias{pw_mat_to_df}
\title{Convert a pairwise matrix into an edge-list data.frame}
\usage{
pw_mat_to_df(pw_mat)
}
\arguments{
\item{pw_mat}{A pairwise matrix which can be:\itemize{
\item{An object of class \code{matrix}. It must have
the same row names and column names. If values represent distances,
diagonal elements should be equal to 0.}
\item{An object of class \code{dist}. In that, its column numbers are
used to create IDs in the resulting data.frame.}
}}
}
\value{
An object of class \code{data.frame}
}
\description{
The function converts a pairwise matrix into an edge-list
data.frame
}
\examples{
data(data_tuto)
pw_mat <- data_tuto[[1]]
df <- pw_mat_to_df(pw_mat)
}
\author{
P. Savary
}
